easy coding -
recap & re-organized about try-catch-finally。

1 不要在finally中return；
2 在try之前获取到锁。
