--- src/main/java/com/henry/tryout/easy_coding/exception_and_log/Hello.java
@@ -1,5 +1,0 @@
-package com.henry.tryout.easy_coding.exception_and_log;
-
-public class Hello {
-
-}
