+++ src/main/java/com/henry/tryout/easy_coding/exception_and_log_05/Hello.java
@@ -1,0 +1,5 @@
+package com.henry.tryout.easy_coding.exception_and_log_05;
+
+public class Hello {
+
+}
