+++ src/main/java/com/henry/tryout/easy_coding/exception_and_log_05/try_block_02/try_block_01
@@ -21,4 +21,4 @@ finally代码块；
 
         3 finally代码块 会在 return表达式执行完后 才执行；
         原理：要return的结果会先被暂存起来，等到 finally代码块执行结束后，再返回 先前暂存的结果。
-    代码 demo: finallyNotWork
+    代码 demo: finallyNotWorkDemo_01
